maybe with 64 bit keys?

+++ src/engine/transposition_table.h
@@ -18,7 +18,7 @@ enum {
 
 typedef uint64_t TTLength_t;
 typedef uint32_t TTIndex_t;
-typedef uint16_t TTKey_t;
+typedef uint64_t TTKey_t;
 
 typedef struct {
     TTFlag_t flag;
